requirements.txt
torch>=1.9.0
transformers>=4.5.0
Pillow>=8.0.0
gradio>=2.0.0
tqdm>=4.62.0
python-dotenv>=0.19.0
psutil>=5.9.0
